feat: add typescript support
