+++ package.json
@@ -2,27 +2,33 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
+    "main": "index.ts",
     "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "dev": "nodemon",
+        "start": "cross-env ts-node index.ts",
+        "build": "cross-env tsc --build&&node dist/index.js",
+        "start:dev": "cross-env nodemon index.ts"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
+        "@types/node": "^18.0.0",
         "@types/ws": "^8.5.3",
         "@typescript-eslint/eslint-plugin": "^5.29.0",
         "@typescript-eslint/parser": "^5.29.0",
+        "cross-env": "^7.0.3",
         "eslint": "^8.18.0",
         "eslint-config-airbnb": "^19.0.4",
         "eslint-config-airbnb-typescript": "^17.0.0",
         "nodemon": "^2.0.16",
+        "ts-loader": "^9.3.1",
+        "ts-node": "^10.8.1",
         "typescript": "^4.7.4"
     },
     "dependencies": {
+        "dotenv": "^16.0.1",
         "jimp": "^0.16.1",
         "robotjs": "^0.6.0",
         "ws": "^8.8.0"
