--- src/actions/mousePosition.js
@@ -1,7 +1,0 @@
-import robot from 'robotjs';
-
-export default (cmdKey, duplex, YDistance) => {
-  const { x, y } = robot.getMousePos();
-
-  duplex.write(`${cmdKey} ${x}px,${y}px \0`);
-};
